backend/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7
pymongo==4.5.0
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
alembic==1.12.0

# Authentication
Flask-JWT-Extended==4.5.2
passlib==1.7.4
bcrypt==4.0.1

# Image and video processing
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-image==0.21.0
numpy==1.25.2
scipy==1.11.2

# Deep learning
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
transformers==4.32.1

# Feature extraction
exifread==3.0.0
pyexiftool==0.5.5

# Utilities
requests==2.31.0
tqdm==4.66.1
pandas==2.1.0
matplotlib==3.7.2
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
isort==5.12.0
